fix: separate EPUB navigation direction from layout
